- updated Scripts - code cleaning
