+++ helper/prompt_templates.py
@@ -8,7 +8,7 @@ Note: If you change the structure not only the semantics, you may need to adjust
 
 """
 
-
+# region General Pipeline Prompts
 prompt_template_translation = PromptTemplate.from_template(
 '''You are a professional translator. Translate the following text into English if it is not already in English.
 
@@ -33,7 +33,6 @@ TEXT: "{text}"
 TRANSLATION:
 '''
 )
-
 
 prompt_template_topic = PromptTemplate.from_template(
 '''
@@ -243,7 +242,6 @@ TOPICS:
 '''
 )
 
-
 prompt_template_sentiment = PromptTemplate.from_template(
 '''What's the sentiment of the review with regard to the topic?
 Always answer with 'Positive' or 'Negative' or 'Inconclusive'.
@@ -265,7 +263,43 @@ TOPIC: {topic}
 SENTIMENT: '''
 )
 
+prompt_template_cluster_naming = (
+'''Based on the following topics, generate a concise name (5 words or fewer) that best describes the general theme of this cluster.
 
+TOPICS: {topics}
+CLUSTER NAME: '''
+)
+
+# endregion
+
+# region Prompt templates for the Cluster Report
+
+# Used for the cluster report (individual clusters)
+prompt_template_summary_short = PromptTemplate.from_template(
+    """
+You are analyzing player statements for our video game {video_game}.
+Below are the player statements from the cluster: {cluster_name}.
+
+Please provide a short summary (2-3 sentences) that captures the main topics of those statements:
+{statements}
+"""
+)
+# Used for the Big Picture Report (top 5 clusters)
+prompt_template_top5 = PromptTemplate.from_template(
+    """
+You are analyzing player statements for our video game {video_game}.
+Below are the player statements from the {cluster_group} clusters.
+
+Your task is to identify the the main {sentiment} topics in the statements. Give a brief summary about them in 2-3 sentences.
+Here are the player statements:
+
+{statements}
+"""
+)
+
+# endregion
+
+# region experimental
 prompt_template_influencer = PromptTemplate.from_template(
 '''
 Please list the most important topics and their respective original context in the excerpt of a video transcript in a JSON format with "Topic", "Category", and "Context" arguments.
@@ -304,72 +338,4 @@ TOPICS:
 '''
 )
 
-prompt_template_summary = PromptTemplate.from_template(
-'''
-cluster_name
-You are analysing player statements for our video game {video_game}. This is the cluster where you should analyse the player statements:
-{sentiment_distribution}
-Provide a concise summary where you name any major themes or issues that stand out.
-List the key insights, concerns or praises in this cluster as **bullet points** below.
-
-Here are the player statements in that cluster:
-
-{statements}
-'''
-)
-
-prompt_template_summary_short = PromptTemplate.from_template(
-    """
-You are analyzing player statements for our video game {video_game}.
-Below are the player statements from the cluster: {cluster_name}.
-
-Please provide a short summary (2-3 sentences) that captures the main topics of those statements:
-{statements}
-"""
-)
-
-prompt_template_pain_points = PromptTemplate.from_template(
-    """
-You are analyzing player statements for our video game {video_game}.
-Below are the player statements from the cluster: {cluster_name}.
-
-Your task is to identify the biggest **pain points** (complaints or issues) mentioned by players.
-Provide them as **bullet points**. Only list frequent or significant pain points (maximum 3).
-If you cant identify any pain points just write "No issues mentioned".
-{statements}
-"""
-)
-
-prompt_template_highlights = PromptTemplate.from_template(
-    """
-You are analyzing player statements for our video game {video_game}.
-Below are the player statements from the cluster: {cluster_name}.
-
-Your task is to identify the **biggest strengths** (praises or well perceived aspects) mentioned by players.
-Provide them as **bullet points**. Only list frequent or significant strengths (maximum 3).
-If you cant identify any highlights just write "No strengths mentioned".
-{statements}
-"""
-)
-
-prompt_template_top5 = PromptTemplate.from_template(
-    """
-You are analyzing player statements for our video game {video_game}.
-Below are the player statements from the {cluster_group} clusters.
-
-Your task is to identify the the main {sentiment} topics in the statements. Give a brief summary about them in 2-3 sentences.
-Here are the player statements:
-
-{statements}
-"""
-)
-
-
-
-
-prompt_template_cluster_naming = (
-'''Based on the following topics, generate a concise name (5 words or fewer) that best describes the general theme of this cluster.
-
-TOPICS: {topics}
-CLUSTER NAME: '''
-)
+# endregion
